Update CORS origin to new local IP
The CORS origin was previously set to 'http://localhost:5173'. However, to align with recent updates to our local development infrastructure, the origin has been updated to 'http://192.168.1.87:3000'.

This ensures seamless interoperability with other services running in the local network. Please ensure that your own environments are updated to reflect this change if necessary.